+++ resources/views/ekskuls/index.blade.php
@@ -66,9 +66,13 @@
                             </td> --}}
                             <td>
                                 <a class="btn btn-info btn-sm" href="/ekskuls/{{ $data->id }}/detail">Detail</a>
-                                <a class="btn btn-warning btn-sm" href="/ekskuls/{{ $data->id }}/edit">Edit</a>
-                                <a class="btn btn-danger btn-sm" href="/ekskuls/{{ $data->id }}/delete"
-                                    onClick="return confirm('Anda Yakin ?')">Delete</a>
+                                @if (Auth::user()->id_role == 1)
+                                    <a class="btn btn-warning btn-sm" href="/ekskuls/{{ $data->id }}/edit">Edit</a>
+                                    <a class="btn btn-danger btn-sm" href="/ekskuls/{{ $data->id }}/delete"
+                                        onClick="return confirm('Anda Yakin ?')">Delete</a>
+                                @else
+                                -   
+                                @endif
                             </td>
                         </tr>
                         @endforeach
