+++ resources/views/teachers/index.blade.php
@@ -60,10 +60,14 @@
                             <td>{{ $data->id }}</td>
                             <td>{{ $data->name }}</td>
                             <td>
-                                <a class="btn btn-info btn-sm" href="/teachers/{{ $data->id }}/detail">Detail</a>
-                                <a class="btn btn-warning btn-sm" href="/teachers/{{ $data->id }}/edit">Edit</a>
-                                <a class="btn btn-danger btn-sm" href="/teachers/{{ $data->id }}/delete"
-                                    onClick="return confirm('Anda Yakin ?')">Delete</a>
+                                @if (Auth::user()->id_role == 1)
+                                    <a class="btn btn-info btn-sm" href="/teachers/{{ $data->id }}/detail">Detail</a>
+                                    <a class="btn btn-warning btn-sm" href="/teachers/{{ $data->id }}/edit">Edit</a>
+                                    <a class="btn btn-danger btn-sm" href="/teachers/{{ $data->id }}/delete"
+                                        onClick="return confirm('Anda Yakin ?')">Delete</a>
+                                @else
+                                -
+                                @endif
                             </td>
                         </tr>
                         @endforeach
